Kotlin multiplatform support #19:  - kotlin common/native generation  - specific code generation for hardware encoding/decoding Android/iOS  - only use signed integer type in common code Support decoder flushing with empty/null IPacket Remove unnecessary dynamic_cast

+++ src/com/xuggle/mediatool/MediaWriter.java
@@ -901,7 +901,7 @@ implements IMediaWriter
 
       // establish the number of samples
 
-      long sampleCount = samples.length / coder.getChannels();
+      int sampleCount = samples.length / coder.getChannels();
 
       // create the audio samples object and extract the internal buffer
       // as an array
